Remove unnecessary includes and remove some nested includes and dependencies.
This commit:
1) Removes some completely unnecessary includes from headers
2) Explicitly added some necessary imports that were only included indirectly
3) Change some types to remove the need of importing large headers (such as world, which drags in more dependencies that f.e. worldDataView)
