+++ src/stp/evaluations/global/GoalVisionFromBallGlobalEvaluation.cpp
@@ -29,7 +29,7 @@ GoalVisionFromBallGlobalEvaluation::GoalVisionFromBallGlobalEvaluation() noexcep
 }
 
 uint8_t GoalVisionFromBallGlobalEvaluation::metricCheck(const world::World* world, const world::Field* field) const noexcept {
-    return calculateMetric(FieldComputations::getPercentageOfGoalVisibleFromPoint(*field, false, world->getWorld()->getBall().value()->getPos(), world, -1, true));
+    return calculateMetric(FieldComputations::getPercentageOfGoalVisibleFromPoint(*field, false, world->getWorld()->getBall().value()->getPos(), world->getWorld().value(), -1, true));
 }
 
 uint8_t GoalVisionFromBallGlobalEvaluation::calculateMetric(const double& x) const noexcept { return piecewiseLinearFunction->yForX(x); }
