+++ src/stp/tactics/KeeperBlockBall.cpp
@@ -7,6 +7,7 @@
 #include "control/ControlUtils.h"
 #include "stp/skills/GoToPos.h"
 #include "stp/skills/Rotate.h"
+#include "roboteam_utils/LineSegment.h"
 
 namespace rtt::ai::stp::tactic {
