+++ src/stp/tactics/active/GetBall.cpp
@@ -11,6 +11,7 @@
 #include "stp/skills/GoToPos.h"
 #include "stp/skills/Rotate.h"
 #include "utilities/GameStateManager.hpp"
+#include "world/FieldComputations.h"
 
 namespace rtt::ai::stp::tactic {
 GetBall::GetBall() { skills = collections::state_machine<Skill, Status, StpInfo>{skill::GoToPos(), skill::Rotate()}; }
